requirements.txt
Flask>=2.0
psycopg2-binary>=2.9
python-dotenv>=0.21
SQLAlchemy>=2.0
SQLAlchemy>=2.0
